+++ libs/angular/src/lib/sortable-list/index.ts
@@ -1,0 +1,2 @@
+export * from './sortable-list.component'
+export * from './sortable-list.module'